requirements-dev.txt
flask 
celery 
gevent # for windows
redis
python-dotenv
pytest # for testing
requests # for testing